cli/pyproject.toml
[project]
name = "castle-cli"
version = "0.1.0"
description = "Castle platform CLI - manage projects, services, and infrastructure"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
]

[project.scripts]
castle = "castle_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/castle_cli"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.11.0",
    "pyright>=1.1.0",
]
